(maint) Add a --dump action to the parser face
This adds the --dump action to the parser face which performs a
tree dump of the produced AST model using the future parser.

The command accepts an additional --[no]-validate option to turn off
validation (the default validates). Without validation only syntax
errors terminate the dump.

The -e option (chosen to mimic -e to apply) accepts the source on the
command line. Alternatively the source can be fed to stdin, or given as
a series of files (without -e option)
This command is useful because it is otherwise required to instrument
the code to get a dump of the result when working on parser related
bugs.